[refactor] creating time filter package
