+++ apolline-flutter/lib/models/user_configuration.dart
@@ -1,22 +1,6 @@
 import 'package:apollineflutter/models/data_point_model.dart';
-import 'package:apollineflutter/services/user_configuration_service.dart';
+import 'package:apollineflutter/utils/time_filter.dart';
 
-
-///Author (Issagha BARRY)
-///
-enum TimeFilter {
-  MAP_SYNC_1_MIN,
-  MAP_SYNC_5_MIN,
-  MAP_SYNC_15_MIN,
-  MAP_SYNC_30_MIN,
-  MAP_SYNC_1_HOUR,
-  MAP_SYNC_3_HOUR,
-  MAP_SYNC_6_HOUR,
-  MAP_SYNC_12_HOUR,
-  MAP_SYNC_24_HOUR,
-  MAP_SYNC_TODAY,
-  MAP_SYNC_THIS_WEEK,
-}
 
 ///Author (Issagha BARRY)
 ///User configuration in ui
