+++ apolline-flutter/lib/services/sqflite_service.dart
@@ -1,9 +1,10 @@
 import 'dart:io';
 import 'package:apollineflutter/models/data_point_model.dart';
+import 'package:apollineflutter/utils/time_filter.dart';
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:apollineflutter/models/user_configuration.dart';
+
 
 // Author GDISSA Ramy
 // Sqflite Database
