+++ apolline-flutter/lib/utils/time_filter.dart
@@ -1,0 +1,13 @@
+enum TimeFilter {
+  MAP_SYNC_1_MIN,
+  MAP_SYNC_5_MIN,
+  MAP_SYNC_15_MIN,
+  MAP_SYNC_30_MIN,
+  MAP_SYNC_1_HOUR,
+  MAP_SYNC_3_HOUR,
+  MAP_SYNC_6_HOUR,
+  MAP_SYNC_12_HOUR,
+  MAP_SYNC_24_HOUR,
+  MAP_SYNC_TODAY,
+  MAP_SYNC_THIS_WEEK,
+}
